much of the conversion of the UserGroupSuggestBox to MVP.
